JokeCard component created working on styling
