+++ src/App/App.scss
@@ -38,11 +38,11 @@
   }
 }
 
-.joke-container {
-  // display: flex;
-  // justify-content: center;
-  // position: absolute;
+.Joke-container {
+  display: flex;
+  justify-content: center;
+  position: absolute;
   height: 735px;
   background: #474747;
-  border-radius: 20px;
+  border-radius: 50px;
 }
